Move token activation button to separate component
