+++ src/data/generated.ts
@@ -1792,7 +1792,7 @@ export type UserBalanceWithLockQuery = { user: (
     Pick<User, 'id'>
     & { userLock: (
       Pick<UserLock, 'balance' | 'totalObligation'>
-      & { nativeToken?: Maybe<Pick<UserToken, 'decimals' | 'name' | 'symbol' | 'balance'>> }
+      & { nativeToken?: Maybe<Pick<UserToken, 'decimals' | 'name' | 'symbol' | 'balance' | 'address' | 'verified'>> }
     ) }
   ) };
 
@@ -3934,6 +3934,8 @@ export const UserBalanceWithLockDocument = gql`
         name
         symbol
         balance
+        address
+        verified
       }
       totalObligation
     }
