+++ src/modules/pages/components/RouteLayouts/UserNavigation.css.d.ts
@@ -11,6 +11,3 @@ export const walletAddress: string;
 export const walletAddressActive: string;
 export const buttonsWrapper: string;
 export const readyTransactionsCount: string;
-export const tokens: string;
-export const dot: string;
-export const dotInactive: string;
