+++ src/modules/users/components/UserTokenActivationButton/UserTokenActivationButton.css.d.ts
@@ -1,0 +1,3 @@
+export const tokens: string;
+export const dot: string;
+export const dotInactive: string;
