+++ src/modules/users/components/UserTokenActivationButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserTokenActivationButton';
